elevatorsystem/requirements.txt
Django==4.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt
django-auditlog
requests
psycopg2
django-rest-framework-nested